Added DevUtils/Nodes and DevUtils/Searches subpackages + fixed errors caused by moving SearchAlgorithm
