+++ Searches/BinarySearch.java
@@ -6,6 +6,7 @@ import java.util.Arrays;
 import java.util.Random;
 import java.util.concurrent.ThreadLocalRandom;
 import java.util.stream.IntStream;
+import DevUtils.Searches.SearchAlgorithm;
 
 /**
  * Binary search is one of the most popular algorithms The algorithm finds the position of a target
