+++ Searches/LowerBound.java
@@ -5,6 +5,7 @@ import static java.lang.String.format;
 import java.util.Random;
 import java.util.concurrent.ThreadLocalRandom;
 import java.util.stream.IntStream;
+import DevUtils.Searches.SearchAlgorithm;
 
 /**
  * The LowerBound method is used to return an index pointing to the first element in the range
