+++ Searches/TernarySearch.java
@@ -5,6 +5,7 @@ import static java.lang.String.format;
 import java.util.Arrays;
 import java.util.Random;
 import java.util.stream.Stream;
+import DevUtils.Searches.SearchAlgorithm;
 
 /**
  * A ternary search algorithm is a technique in computer science for finding the minimum or maximum
